Git Update : - Improvment to be compliant with REST API - New fields to add pet's information

+++ petshop-ws/src/main/java/my/petshop/service/PetShopService.java
@@ -3,6 +3,8 @@ package my.petshop.service;
 
 import java.util.ArrayList;
 
+import javax.servlet.http.HttpServletResponse;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -24,38 +26,54 @@ public class PetShopService {
      * POST /create  --> Create a new pet and save it in the database.
      */
     @RequestMapping(value = "/pet", method= RequestMethod.POST)
-    public String create(@RequestBody PetHolder pPetHolder) {
+    public String create(@RequestBody PetHolder pPetHolder, HttpServletResponse response) {
     
     	if (pPetHolder == null ||
     			pPetHolder.getList().isEmpty() ||
     			pPetHolder.getList().get(0).getClass().isInstance(PetEntity.class)){
-    		return "ERROR DURING CREATION";
+    		//Will return HTTP Code 405
+    		response.setStatus(HttpServletResponse.SC_METHOD_NOT_ALLOWED);
+    		System.out.println("Invalid input ["+pPetHolder.getList()+"]");
+    		return "KO";
     	}
     		
       try {
         PetEntity pet = pPetHolder.getList().get(0);
         pet = petDao.save(pet);
-      }
-      catch (Exception ex) {
-    	  ex.printStackTrace();
-        return "Error creating the pet: " + ex.toString();
-      }
-      return "Pet succesfully created";
+    } catch (Exception ex) {
+	  	  //Will return HTTP Code 500
+	  	  response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+	  	  System.out.println("Exception raised");
+	  	  ex.printStackTrace();
+    }
+      return "OK";
     }
 
     /**
      * DELETE /delete  --> Delete the pet having the passed id.
      */
     @RequestMapping(value = "/pet/{id}", method= RequestMethod.DELETE)
-    public String delete(@PathVariable("id") String id) {
+    public String delete(@PathVariable("id") String id, HttpServletResponse response) {
       try {
-    	PetEntity pet = new PetEntity(Long.valueOf(id));
-        petDao.delete(pet);
+    	PetEntity pet = petDao.findById(Long.valueOf(id));
+    	if (pet == null){
+        	//Will return HTTP Code 404
+    		response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+    		System.out.println("No pet found with id ["+id+"]");
+    	} else {
+    		petDao.delete(pet);
+    	}
+      }catch (NumberFormatException nfe) {
+    	//Will return HTTP Code 400
+    	response.setStatus(HttpServletResponse.SC_BAD_REQUEST );
+    	System.out.println("NumberFormatException ["+id+"] is not a number");
+      } catch (Exception ex) {
+    	  //Will return HTTP Code 500
+    	  response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+    	  System.out.println("Exception raised ["+id+"]");
+    	  ex.printStackTrace();
       }
-      catch (Exception ex) {
-        return "Error deleting the pet:" + ex.toString();
-      }
-      return "Pet succesfully deleted!";
+      return "OK";
     }
     
     /**
@@ -63,7 +81,7 @@ public class PetShopService {
      * In case there is no id, the service returns all the pets
      */
     @RequestMapping(value = "/pet/{id}", method= RequestMethod.GET)
-    public PetHolder getById(@PathVariable("id") String id) {
+    public PetHolder getById(@PathVariable("id") String id, HttpServletResponse response) {
     	
     	PetHolder output = new PetHolder();
     	try {
@@ -72,11 +90,21 @@ public class PetShopService {
     	}else {
     		output.add(petDao.findById(Long.valueOf(id)));
     	}
-      }
-      catch (Exception ex) {
+    	if (output.getList().isEmpty() || output.getList().get(0) == null){
+        	//Will return HTTP Code 404
+    		response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+    		System.out.println("No pet found with id ["+id+"]");
+    	}
+      }catch (NumberFormatException nfe) {
+    	//Will return HTTP Code 400
+    	response.setStatus(HttpServletResponse.SC_BAD_REQUEST );
+    	System.out.println("NumberFormatException ["+id+"] is not a number");
+      } catch (Exception ex) {
+    	  //Will return HTTP Code 500
+    	  response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+    	  System.out.println("Exception raised ["+id+"]");
     	  ex.printStackTrace();
       }
-
       return output;
     }
 }
